Add stateless Legacy service wrappers for factory-injectable interface implementations
