+++ src/Rest_API/Barcode/Client.php
@@ -1,53 +1,19 @@
 <?php
 /**
- * Class Rest_API\Barcode\Client file.
+ * Backward-compatibility shim for Rest_API\Barcode\Client.
+ *
+ * The implementation has moved to Rest_API\Legacy\Barcode\Client.
+ * This file registers a class alias so all existing callers that reference
+ * PostNLWooCommerce\Rest_API\Barcode\Client continue to work unchanged.
  *
  * @package PostNLWooCommerce\Rest_API\Barcode
  */
-
-namespace PostNLWooCommerce\Rest_API\Barcode;
-
-use PostNLWooCommerce\Rest_API\Base;
-use PostNLWooCommerce\Utils;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-/**
- * Class Client
- *
- * @package PostNLWooCommerce\Rest_API\Barcode
- */
-class Client extends Base {
-	/**
-	 * API Endpoint.
-	 *
-	 * @var string
-	 */
-	public $endpoint = '/shipment/v1_1/barcode';
-
-	/**
-	 * PostnL API Method.
-	 *
-	 * @var string
-	 */
-	public $method = 'GET';
-
-	/**
-	 * Function for composing API request in the URL for GET request.
-	 *
-	 * @return Array.
-	 */
-	public function compose_url_params() {
-		$range = Utils::get_barcode_range( $this->item_info->query_args['barcode_type'], $this->item_info->query_args['globalpack_customer_code'] );
-
-		return array(
-			'Type'           => $this->item_info->query_args['barcode_type'],
-			'Serie'          => $this->item_info->query_args['serie'],
-			'CustomerCode'   => $this->item_info->query_args['customer_code'],
-			'CustomerNumber' => $this->item_info->query_args['customer_num'],
-			'Range'          => $range,
-		);
-	}
-}
+class_alias(
+	'PostNLWooCommerce\\Rest_API\\Legacy\\Barcode\\Client',
+	'PostNLWooCommerce\\Rest_API\\Barcode\\Client'
+);
