+++ src/Rest_API/Legacy/Letterbox/Client.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Class Rest_API\Letterbox\Client file.
+ *
+ * @package PostNLWooCommerce\Rest_API\Legacy\Letterbox
+ */
+
+namespace PostNLWooCommerce\Rest_API\Legacy\Letterbox;
+
+use PostNLWooCommerce\Rest_API\Legacy\Shipping;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Class Client
+ *
+ * @package PostNLWooCommerce\Rest_API\Legacy\Letterbox
+ */
+class Client extends Shipping\Client {
+
+}
