+++ src/Rest_API/Letterbox/Item_Info.php
@@ -1,39 +1,19 @@
 <?php
 /**
- * Class Rest_API\Letterbox\Item_Info file.
+ * Backward-compatibility shim for Rest_API\Letterbox\Item_Info.
+ *
+ * The implementation has moved to Rest_API\Legacy\Letterbox\Item_Info.
+ * This file registers a class alias so all existing callers that reference
+ * PostNLWooCommerce\Rest_API\Letterbox\Item_Info continue to work unchanged.
  *
  * @package PostNLWooCommerce\Rest_API\Letterbox
  */
-
-namespace PostNLWooCommerce\Rest_API\Letterbox;
-
-use PostNLWooCommerce\Rest_API\Shipping;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-/**
- * Class Item_Info
- *
- * @package PostNLWooCommerce\Rest_API\Letterbox
- */
-class Item_Info extends Shipping\Item_Info {
-	/**
-	 * Get product code from api args.
-	 *
-	 * @return String.
-	 */
-	public function get_product_code() {
-		return '2928';
-	}
-
-	/**
-	 * Get product options from api args.
-	 *
-	 * @return String.
-	 */
-	public function get_product_options() {
-		return array();
-	}
-}
+class_alias(
+	'PostNLWooCommerce\\Rest_API\\Legacy\\Letterbox\\Item_Info',
+	'PostNLWooCommerce\\Rest_API\\Letterbox\\Item_Info'
+);
